26_commit
Refactor/Improved endpoints for Cart, Product, and Room
- Standardized response formats/type to JSON for consistency
- Implemented additional error handling mechanisms

+++ src/main/java/com/example/ApartmentRenovationCostEstimate/cart/CartController.java
@@ -1,15 +1,19 @@
 package com.example.ApartmentRenovationCostEstimate.cart;
 
+
+import com.example.ApartmentRenovationCostEstimate.response.ApiResponse;
+import com.example.ApartmentRenovationCostEstimate.response.ErrorResponse;
 import com.example.ApartmentRenovationCostEstimate.product.Product;
+import com.example.ApartmentRenovationCostEstimate.response.ErrorType;
 import com.example.ApartmentRenovationCostEstimate.user.User;
 import com.example.ApartmentRenovationCostEstimate.product.ProductService;
 import com.example.ApartmentRenovationCostEstimate.user.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -33,28 +37,30 @@ public class CartController {
     @PostMapping("/create")
     public ResponseEntity<Object> createCart(@RequestBody Map<String, Object> payload){
         if (!payload.containsKey("userId") || payload.get("userId") == null) {
-            return new ResponseEntity<>("userId is missing", HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new ErrorResponse(ErrorType.USERID_IS_MISSING), HttpStatus.BAD_REQUEST);
         }
         if (!payload.containsKey("name") || payload.get("name") == null) {
-            return new ResponseEntity<>("name is missing", HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new ErrorResponse(ErrorType.CART_NAME_IS_MISSING), HttpStatus.BAD_REQUEST);
         }
+
         Long userId;
         String name;
+
         try {
              userId = Long.valueOf(payload.get("userId").toString());
              name = payload.get("name").toString();
         } catch (NumberFormatException e) {
-            return new ResponseEntity<>("Invalid userId format", HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(new ErrorResponse(ErrorType.INVALID_USER_FORMAT), HttpStatus.BAD_REQUEST);
         }
 
         Optional<User> userOptional = userService.findById(userId);
         if (userOptional.isEmpty()) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(new ErrorResponse(ErrorType.USER_NOT_FOUND), HttpStatus.NOT_FOUND);
         }
 
         User user = userOptional.get();
         Cart savedCart = cartService.createCart(user, name);
-        return new ResponseEntity<>(savedCart, HttpStatus.CREATED);
+        return new ResponseEntity<>(new ApiResponse<>("Cart created successfully.", savedCart), HttpStatus.CREATED);
     }
 
     @PostMapping("{cartId}/products")
@@ -63,28 +69,31 @@ public class CartController {
         /*Sprawdzanie czy koszyk instnieje*/
         Cart cart = cartService.getCartById(cartId);
         if(cart == null) {
-            return new ResponseEntity<>(Map.of("message", "Cart not found"), HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(new ErrorResponse(ErrorType.CART_NOT_FOUND), HttpStatus.NOT_FOUND);
         }
 
         /*Sprawdzanie czy produkt instnieje*/
         Product product = productService.getProductById(request.getProductId());
         if(product == null) {
-            return new ResponseEntity<>(Map.of("message", "Product not found"), HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(new ErrorResponse(ErrorType.PRODUCT_NOT_FOUND), HttpStatus.NOT_FOUND);
         }
 
         /*Dodanie produktu do koszyka*/
         try {
-            cartService.addProductToCart(cartId, request.getProductId(), request.getQuantity());
-            return new ResponseEntity<>(Map.of("message", "Product added to cart"), HttpStatus.OK);
-        } catch (Exception e) {
-            return new ResponseEntity<>(Map.of("message", "Failed to add product to cart"), HttpStatus.INTERNAL_SERVER_ERROR);
+            Cart addedProduct = cartService.addProductToCart(cartId, request.getProductId(), request.getQuantity());
+            return new ResponseEntity<>(new ApiResponse<>("Product added to cart", addedProduct), HttpStatus.OK);
+        } catch (RuntimeException e) {
+            return new ResponseEntity<>(new ErrorResponse(ErrorType.FAILED_TO_ADD_PRODUCT_TO_CART," An unexpected error occurred"), HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @GetMapping("{cartId}")
-    public ResponseEntity<Cart> getCartById(@PathVariable("cartId") Long cartId) {
+    public ResponseEntity<Object> getCartById(@PathVariable("cartId") Long cartId) {
         Cart cart = cartService.getCartById(cartId);
-        return new ResponseEntity<>(cart, HttpStatus.OK);
+        if(cart == null) {
+            return new ResponseEntity<>(new ErrorResponse(ErrorType.CART_NOT_FOUND), HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(new ApiResponse<>("Cart retrieved successfully",cart), HttpStatus.OK);
     }
 
     @GetMapping()
@@ -94,15 +103,23 @@ public class CartController {
     }
 
     @DeleteMapping("{cartId}/products/{productId}")
-    public ResponseEntity<String> removeProductFromCart(@PathVariable("cartId") Long cartId, @PathVariable("productId") Long productId) {
-        cartService.removeProductFromCart(cartId, productId);
-        return new ResponseEntity<>("Product sucessfully removed", HttpStatus.OK);
+    public ResponseEntity<Object> removeProductFromCart(@PathVariable("cartId") Long cartId, @PathVariable("productId") Long productId) {
+        try {
+            cartService.removeProductFromCart(cartId, productId);
+            return new ResponseEntity<>(new ApiResponse<>("Product successfully removed"), HttpStatus.OK);
+        } catch (ResourceNotFoundException e) {
+            return new ResponseEntity<>(new ErrorResponse(ErrorType.CART_OR_PRODUCT_NOT_FOUND), HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("{cartId}")
-    public ResponseEntity<String> deleteCart(@PathVariable("cartId") Long cartId) {
+    public ResponseEntity<Object> deleteCart(@PathVariable("cartId") Long cartId) {
+        Optional<Cart> cart = cartService.findById(cartId);
+        if (cart.isEmpty()) {
+            return new ResponseEntity<>(new ErrorResponse(ErrorType.CART_NOT_FOUND), HttpStatus.NOT_FOUND);
+        }
         cartService.deleteCart(cartId);
-        return new ResponseEntity<>("Cart successfully deleted", HttpStatus.OK);
+        return new ResponseEntity<>(new ApiResponse<>("Cart successfully deleted"), HttpStatus.OK);
     }
 
     @GetMapping("{cartId}/products/categories")
@@ -111,22 +128,16 @@ public class CartController {
         return new ResponseEntity<>(categories, HttpStatus.OK);
     }
 
+/*    @GetMapping("{cartId}/products/categories")
+    public ResponseEntity<Object> getProductCategoriesFromCart(@PathVariable("cartId") Long cartId) {
+        List<String> categories = cartService.getAllProductCategoriesFromCart(cartId);
+        return new ResponseEntity<>(new ApiResponse<>("Product categories fetched successfully", categories), HttpStatus.OK);
+    }*/
+
     @GetMapping("{cartId}/products/category/{category}")
     public ResponseEntity<List<CartItem>> getProductByCategoryFromCart(@PathVariable("cartId") Long cartId, @PathVariable("category") String category) {
         List<CartItem> productByCategory = cartService.getProductByCategoryFromCart(cartId, category);
         return new ResponseEntity<>(productByCategory, HttpStatus.OK);
     }
-
-    //!!!!! Nieużywane -> zmiana koncepcji.
-    // Wartosc koszyka jest aktualizowana automatycznie w trakcjie dodawania lub usuwania elementu w koszyku!!!!!!!!
-    @GetMapping("{cartId}/cart-total-cost")
-    public ResponseEntity<BigDecimal> calculateTotalCost(@PathVariable Long cartId) {
-        BigDecimal totalCostCart = cartService.calculateCartTotalCost(cartId);
-        if (totalCostCart == null || totalCostCart.compareTo(BigDecimal.ZERO) == 0) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-        return new ResponseEntity<>(totalCostCart, HttpStatus.OK);
-    }
-
 
 }
